Convert scopes to TypeScript
